Remove DRY rules from ruleset
Originally, I wanted to suppress the DuplicateListLiteral rule to this
set, but then I realized that we are basically suppressing the entire
ruleset in this case. These rules are not that interesting to us,
primarily because they cannot account for scope, which means that
fixing violations thrown by this ruleset is usually done by
introducing global variables/fields, which is arguably worse.