Pull Docker image before withRun()
The documentation claims that explicitly calling pull() isn't
necessary, which is partially true because run/withRun/inside will
check to see that the image exists and call pull() if needed. However,
these functions only check that the image is present, not if a newer
one exists on the server. For that reason, we must call pull()
explicitly, since we overwrite tags on the server for major/minor
versions.