Return XML report from run_codenarc()
This change modifies run_codenarc() to return a string with the
contents of the XML report. This function should not be responsible
for parsing the report text itself. By narrowing the responsibility of
this function, it will make it easier to test.

Also, this function has been modified to raise an error instead of
returning an integer code. We should not return mixed types from
functions.