Add script for downloading JAR dependencies
This Python script facilitates the downloading of the JAR file
dependencies needed by run_codenarc.py. By moving this work out of
Dockerfile RUN commands and into a separate script, developers can
also run the script locally and then use this value for
GROOVYLINT_HOME with run_codenarc.py.

This is therefore a first step on the way to having a run_codenarc.py
file that can be run from outside of the Docker container.