Introduce logging support to run_codenarc.py
This change adds `--verbose` and `--quiet` arguments to the script,
which controls the amount of logging produced. We now also parse the
logging produced by CodeNarc and re-log it to the logging framework
using the same log level. This way, the verbosity arguments can also
be used to control how much output CodeNarc itself produces.