Use ElementTree for XML parsing
The xmltodict package was not really great. Accessing the children of
an element was a bit dangerous since we could not be sure whether they
would be packaged in a list or not. Also, this library is a third-
party dependency and thus required pipenv to actually run the script.

That said, we still need pipenv because of third-party package
requirements for the fetch_jars.py script.