Add YAML file for specifying CodeNarc dependencies
This commit introduces a simple YAML file to specify the dependencies
which will be downloaded by this script. Sections for types of
dependencies (tarballs, JAR files) contain entries, each containing a
version number and a Jinja2-templated URL. The fetch_jars.py script
processes the templates and determines the URLs to download.

Note that the Dockerfile still contains the version numbers for the
dependencies as environment variables. This is because run_codenarc.py
still needs these variables in order to determine the version numbers
of the JAR files. A future commit will change the behavior of this
script to make these values obsolete.