Find dependencies in the Groovylint home directory
This commit adds a function to determine the path of a given
dependency. If the version number is given, then an exact filename
will be used. Otherwise, a glob will be used to find the dependency
in the given search directory.

Ideally, we'd like to use the same YAML file as the ‘fetch_jars.py‘
script to specify dependencies, but that would require adding a
third-party import to this script.

This makes the --*-version arguments optional, but not obsolete. These
options are still useful for specifying an exact version of a
dependency, which can be used to test newer versions of CodeNarc or
other dependencies.

With this commit, calling run_codenarc.py is now much more concise,
and in the standard use-case only the --home directory is required.