Remove tarballs after extraction
If we leave the tarballs in the output_dir after extraction, then it
will cause the globbing feature of run_codenarc.py to fail when it
encounters multiple results for these files.

The downside of this is that when running fetch_jars.py, the tarballs
will not be skipped if they exist. Given how often this script is run,
that's a small consequence, though.