Use docker.image.inside instead of runWith
I suspect something has change with the behavior of this function in a
newer release of the docker-workflow plugin, since this singleton
hasn't changed in ages. Anyways, if the entrypoint command fails in
`withRun()`, it no longer causes the pipeline itself to fail.
Furthermore, the closure body is still executed!

Fortunately, we can kill two birds with a single stone here. By using
`image.inside()` (which for reasons that I can no longer recall gave
us trouble in previous versions of docker-workflow), the code will
fail correctly and we can also remove the hacky wait/logs commands
which were previously necessary to make this function work correctly.