Download SLF4J from Maven Central
This change was necessitated by the fact that SLF4J binaries are no
longer hosted on the slf4j.org website, which also no longer offers
HTTPS downloads (as seen in c2650ec).

However, this allows us to kill two birds with one stone, as Maven
doesn't offer a combined tarball but rather individual JAR files. So
instead we can just download the individual JARs that we need and
remove all of the logic related to uncompressing tarballs.