Update CodeNarc to version 1.5
The CompileStatic rule is being suppressed because it doesn't make
much sense in Jenkinsfiles, and is difficult to suppress there because
the Jenkinsfile itself is a class which inherits from WorkflowScript.

We are now using GitHub to download CodeNarc releases, since as of
version 1.4 they have started migrating releases there. Version 1.5
wasn't available on SourceForge as of this commit's date.