Add explicit isort dependency
We use Dependabot for updating dependencies in this repo, but it only
considers dependencies that are listed in the requirements-*.in files.
As flake8-isort is listed, but isort (which the former uses under the
hood) is not, it's possible for flake8-isort to be up-to-date, but
actually calling an out-of-date version of isort.

Adding isort to requirements-dev.in means Dependabot will keep this
dependency updated too.