Allow unsafe packages when compiling depdencies
pip-compile considers several packages "unsafe" for pinning. However, we
would like to be able to pin the pip version itself in requirements-dev
("pip" is one of these unsafe packages).

Adding "--allow-unsafe" tells pip-compile to pin versions of "unsafe"
packages. Despite the warning, there's no evidence of any unsafe effects
from this, and pip-tools is considering deprecating this flag and making
it the default behaviour anyway [1].

I've tested this locally and it seems to work without issue.

[1]: https://github.com/jazzband/pip-tools/issues/806#issuecomment-493591664