Add argument --single-file
It's a pretty common use-case (at least for us) that groovylint is
used to lint a single file, usually the Jenkinsfile in a repository.
However, CodeNarc currently scans the entire working directory on
startup, which can cause performance issues with larger repositories:

https://github.com/CodeNarc/CodeNarc/issues/530

This option creates a temporary directory, copies the desired file to
it, and runs CodeNarc there instead. Aside from being faster with
large repositories (as mentioned), this provides a slightly less
clunky command-line syntax, as it does not require "--" and
CodeNarc-specific arguments (which are not listed in "--help").