Don't accidentally create a folder owned by root
In certain repositories, we don't run groovylint on a Git working copy
but on a copy of the Jenkinsfile that was downloaded from GitHub to the
controller. That file is downloaded in `env.WORKSPACE_TMP`, and
`env.WORKSPACE` doesn't exist since it isn't used. Because it doesn't
exist, mounting `env.WORKSPACE` in the Docker container creates it on
the host machine, but owned by root (even though the Docker Pipeline
Plugin passes `-u <user-id>:<group-id>` to `docker run`).

In that specific case, `-v ${env.WORKSPACE}:/ws` is also pointless,
since the Jenkinsfile is inside `env.WORKSPACE_TMP`. It still works
because the Docker Pipeline Plugin passes `-w <working-directory>` and
`-v <working-directory>:<working-directory>` to `docker run`.

`-v ${env.WORKSPACE}:/ws` is also pointless when the current working
directory is not `env.WORKSPACE` itself (for instance when using the
`dir` pipeline step).