Write report files in the temporary directory
This fixes an error (for real) where CodeNarc would fail with an
exception if the path contained characters that caused command-line
parsing errors (namely `=`, but there are probably other ones as well).

63c21371 attempted to fix this, but incorrectly (and was subsequently
reverted in b7050a8). CodeNarc needs an absolute path for the report
file or else it won't generate one. By using the temporary directory, we
can ensure that the file will be written and any special characters
won't interfere with CodeNarc's command-line parsing.