Add -failOnError=true to CodeNarc arguments
This argument was added in CodeNarc 3.0.0, but we have been a bit late
in adopting it. Anyways, should CodeNarc throw an exception while
parsing code, we want groovylint to exit with an error rather than
continue executing. In some cases, CodeNarc exceptions could be red
herrings, but in most cases they are either indicative of faulty code or
a CodeNarc bug, which should be reported and fixed.