Auto-detect Groovy version by running binary
This change makes the `--groovy4` argument obsolete and changes that
argument to be hidden. This is not only more convenient for users, but
it will be necessary for migrating the Docker container to Groovy 4
(which will be done in a later commit).

As this means the `--groovy4` flag is obsolete (although not technically
eliminated), a minor version bump will follow.